SEC-268 moved pdf functions to seperate file
